requirements.txt
Flask==3.0.0
flask-cors==4.0.0
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.1.0
requests==2.31.0
firebase-admin==6.3.0
gunicorn==21.2.0
